+++ change-log.md
@@ -4,6 +4,8 @@
 
 - Add the component tag to extract connected components of a given graph
 - Add the process_str_for_html filter
+- Make the map tag more flexible: make it possible to map selected columns
+- Fix the quotes bug: keep the cell's original format
 
 ## v0.0.1